course-1/Lab5/requirements.txt
strands-agents
strands-agents-tools 
strands-agents-builder
strands-agents[anthropic]
boto3
mcp[cli]
mcp-server-git
uvicorn>=0.27.0
requests
strands-agents[a2a]